requirements.txt
discord.py==2.3.2
gspread
oauth2client
python-dotenv
Quart
hypercorn